# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentStatusUpdate": {
        "properties": {
          "status": {
            "pattern": "^(active|suspended)$",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "AgentStatusUpdate",
        "type": "object"
      },
...
      "CampusCreate": {
        "properties": {
          "city": {
            "title": "City",
            "type": "string"
          },
          "feature_flags": {
            "additionalProperties": true,
            "default": {},
            "title": "Feature Flags",
            "type": "object"
          },
          "hero_headline": {
            "title": "Hero Headline",
            "type": "string"
          },
          "hero_subtext": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hero Subtext"
          },
          "name": {
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "primary_color": {
            "default": "#000000",
            "title": "Primary Color",
            "type": "string"
          },
          "slug": {
            "minLength": 2,
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "default": "active",
            "pattern": "^(active|coming_soon|suspended)$",
            "title": "Status",
            "type": "string"
          },
          "whatsapp_number": {
            "title": "Whatsapp Number",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "city",
          "hero_headline",
          "whatsapp_number"
        ],
        "title": "CampusCreate",
        "type": "object"
      },
      "CampusRead": {
        "properties": {
          "city": {
            "title": "City",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "feature_flags": {
            "additionalProperties": true,
            "title": "Feature Flags",
            "type": "object"
          },
          "hero_headline": {
            "title": "Hero Headline",
            "type": "string"
          },
          "hero_subtext": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hero Subtext"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "primary_color": {
            "title": "Primary Color",
            "type": "string"
          },
          "region_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Id"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "whatsapp_number": {
            "title": "Whatsapp Number",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "city",
          "hero_headline",
          "whatsapp_number",
          "primary_color",
          "status",
          "created_at"
        ],
        "title": "CampusRead",
        "type": "object"
      },
      "CampusStatusUpdate": {
        "properties": {
          "status": {
            "pattern": "^(active|coming_soon|suspended)$",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "CampusStatusUpdate",
        "type": "object"
      },
      "CampusUpdate": {
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "feature_flags": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Flags"
          },
          "hero_headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hero Headline"
          },
          "hero_subtext": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hero Subtext"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "primary_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Color"
          },
          "status": {
            "anyOf": [
              {
                "pattern": "^(active|coming_soon|suspended)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "whatsapp_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Whatsapp Number"
          }
        },
        "title": "CampusUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListingTransferRequest": {
        "properties": {
          "listing_id": {
            "title": "Listing Id",
            "type": "string"
          },
          "new_agent_id": {
            "title": "New Agent Id",
            "type": "string"
          }
        },
        "required": [
          "listing_id",
          "new_agent_id"
        ],
        "title": "ListingTransferRequest",
        "type": "object"
      },
...
      "ManagerAssign": {
        "properties": {
          "managed_campus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Managed Campus Id"
          },
          "managed_region_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Managed Region Id"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "ManagerAssign",
        "type": "object"
      },
      "ManagerRead": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "managed_campus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Managed Campus Id"
          },
          "managed_region_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Managed Region Id"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role"
        ],
        "title": "ManagerRead",
        "type": "object"
      },
...
      "PlatformStats": {
        "properties": {
          "active_listings": {
            "title": "Active Listings",
            "type": "integer"
          },
          "total_agents": {
            "title": "Total Agents",
            "type": "integer"
          },
          "total_listings": {
            "title": "Total Listings",
            "type": "integer"
          },
          "total_students": {
            "title": "Total Students",
            "type": "integer"
          }
        },
        "required": [
          "total_listings",
          "active_listings",
          "total_agents",
          "total_students"
        ],
        "title": "PlatformStats",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__features__agents__schemas__AgentRead": {
        "properties": {
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "campus_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campus Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_featured": {
            "default": false,
            "title": "Is Featured",
            "type": "boolean"
          },
          "is_founder": {
            "default": false,
            "title": "Is Founder",
            "type": "boolean"
          },
          "is_support": {
            "default": false,
            "title": "Is Support",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "portfolio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Url"
          },
          "profile_photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Photo Url"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "whatsapp": {
            "title": "Whatsapp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "whatsapp",
          "status",
          "created_at"
        ],
        "title": "AgentRead",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Rumia Backend API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/agents/transfer": {
      "post": {
        "description": "Reassign hostel listing ownership to a different agent. Admin only.",
        "operationId": "transfer_listing_api_v1_admin_agents_transfer_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Transfer Listing Api V1 Admin Agents Transfer Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transfer Listing Ownership",
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/agents/{agent_id}/status": {
      "patch": {
        "description": "Suspend or activate an agent account. Admin only.",
        "operationId": "update_agent_status_api_v1_admin_agents__agent_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__features__agents__schemas__AgentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Agent Status",
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/campuses": {
      "get": {
        "description": "Fetch all campuses with full management detail. Admin only.",
        "operationId": "list_campuses_admin_api_v1_admin_campuses_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CampusRead"
                  },
                  "title": "Response List Campuses Admin Api V1 Admin Campuses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List All Campuses (Admin)",
        "tags": [
          "Admin Operations"
        ]
      },
      "post": {
        "description": "Create a new university/college campus. Admin only.",
        "operationId": "create_campus_api_v1_admin_campuses_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampusCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampusRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Campus",
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/campuses/{campus_id}": {
      "patch": {
        "description": "Update campus settings or color scheme. Admin only.",
        "operationId": "update_campus_api_v1_admin_campuses__campus_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "campus_id",
            "required": true,
            "schema": {
              "title": "Campus Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampusRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Campus",
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/campuses/{campus_id}/status": {
      "patch": {
        "description": "Suspend or activate a campus. Admin only.",
        "operationId": "update_campus_status_api_v1_admin_campuses__campus_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "campus_id",
            "required": true,
            "schema": {
              "title": "Campus Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampusStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampusRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Campus Status",
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/managers": {
      "get": {
        "description": "Fetch all assigned campus managers. Admin only.",
        "operationId": "list_managers_api_v1_admin_managers_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ManagerRead"
                  },
                  "title": "Response List Managers Api V1 Admin Managers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Managers",
        "tags": [
          "Admin Operations"
        ]
      },
      "post": {
        "description": "Grant campus/region manager role to a user profile. Admin only.",
        "operationId": "assign_manager_api_v1_admin_managers_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagerAssign"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagerRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Campus Manager",
        "tags": [
          "Admin Operations"
        ]
      }
    },
    "/api/v1/admin/stats": {
      "get": {
        "description": "Fetch total counts of listings, active hostels, agents, and students. Admin only.",
        "operationId": "get_stats_api_v1_admin_stats_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformStats"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Platform Stats",
        "tags": [
          "Admin Operations"
        ]
      }
    },
...
    "/api/v1/campuses": {
      "get": {
        "description": "Retrieve all registered campuses (filtered by active status by default).",
        "operationId": "list_campuses_api_v1_campuses_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "active",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CampusRead"
                  },
                  "title": "Response List Campuses Api V1 Campuses Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Campuses",
        "tags": [
          "Campuses"
        ]
      }
    },
    "/api/v1/campuses/{slug}": {
      "get": {
        "description": "Retrieve a single campus by its unique slug identifier (e.g. 'dekut').",
        "operationId": "get_campus_api_v1_campuses__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampusRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Campus Details",
        "tags": [
          "Campuses"
        ]
      }
    },
...
  }
}